% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acceptable_augment_colnames}
\alias{acceptable_augment_colnames}
\title{Determine acceptable names for augment output}
\usage{
acceptable_augment_colnames(object, passed_data)
}
\arguments{
\item{object}{A model object.}

\item{passed_data}{The dataset used to create the model object.}
}
\value{
A vector of colnames that are acceptable in augment output.
}
\description{
Given a data frame (or tibble), and a model object, makes a
character vector of acceptable columns names for augment output. This
includes:
\itemize{
\item Any column names of the passed dataset
\item Any syntactically correct column names generated by calling
\code{\link[stats:model.frame]{stats::model.frame()}} on the object in question.
}
}
\examples{

fit <- lm(hp ~ log(mpg), mtcars)
acceptable_augment_colnames(fit, mtcars)

}
