% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{accuracy}
\alias{accuracy}
\title{Calculate accuracy}
\usage{
accuracy(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
numeric value
}
\description{
This function calcuates the fraction of the time that the classifier
correctly identifies true and false values.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
accuracy(t, p)
}

