% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_arguments}
\alias{check_arguments}
\title{Check that tidying methods use allowed argument names}
\usage{
check_arguments(tidy_method, strict = FALSE)
}
\arguments{
\item{tidy_method}{A tidying method. For example: \code{glance.Arima}.}
}
\description{
Use this function to test tidying methods in broom tests.
Called for side effects. Will throw a useful error if
\code{tidy_method} has unacceptably named arguments.
Otherwise silent.
}
\examples{

library(broom)
check_arguments(tidy.Arima)

}
\seealso{
\link{testthat}, \code{\link[testthat:expect_true]{testthat::expect_true()}}
}
