% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_augment}
\alias{check_augment}
\title{check the output of an augment function}
\usage{
check_augment(au, original = NULL, exp.names = NULL, same = NULL)
}
\description{
check the output of an augment function
}
