% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_augment_data_specification}
\alias{check_augment_data_specification}
\title{Title}
\usage{
check_augment_data_specification(aug, model, data, add_missing,
  test_newdata)
}
\arguments{
\item{test_newdata}{}
}
\description{
Title
}
