% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_augment_function}
\alias{check_augment_function}
\title{Check that an augment method works correctly}
\usage{
check_augment_function(aug, model, data = NULL, newdata = NULL,
  strict = FALSE)
}
\arguments{
\item{newdata}{}
}
\description{
Check that an augment method works correctly
}
\examples{

library(betareg)
fit <- betareg(yield ~ batch + temp, data = GasolineYield)

check_augment_function(
  aug = augment.betareg,
  model = fit,
  data = GasolineYield,
  newdata = GasolineYield
)

}
