% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_augment_newdata_precedence.R
\name{check_augment_newdata_precedence}
\alias{check_augment_newdata_precedence}
\title{Check that newdata argument has higher precedence than data argument}
\usage{
check_augment_newdata_precedence(aug, model, data, strict = TRUE)
}
\arguments{
\item{aug}{An augment method. For example, \code{augment.betareg}.}

\item{model}{A fit model object to call the augment method on.}

\item{data}{A data frame or tibble to use when testing \code{aug}.}

\item{strict}{Logical indicating whether the strict version of tests should be used. Defaults
to \code{TRUE}.}
}
\value{
An invisible \code{NULL}. This function should be called for side effects, not return values.
}
\description{
Call this function to perform tests. If a tests fails, an informative error
will be thrown. Otherwise silent.
}
