% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_dims}
\alias{check_dims}
\title{Check that tibble has expected dimensions.}
\usage{
check_dims(data, expected_rows = NULL, expected_cols = NULL)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{expected_rows}{Expected number of rows of tibble.}

\item{expected_cols}{Expected number of columns of tibble.}

\item{strict}{Logical indicating whether the strict version of tests should be used. Defaults
to \code{TRUE}.}
}
\value{
An invisible \code{NULL}. This function should be called for side effects, not return values.
}
\description{
Call this function to perform tests. If a tests fails, an informative error
will be thrown. Otherwise silent.
}
\examples{

check_dims(iris, expected_rows = 150)

}
