% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_glance_outputs}
\alias{check_glance_outputs}
\title{Check the output of a glance method}
\usage{
check_glance_outputs(..., strict = FALSE)
}
\arguments{
\item{...}{Outputs returned from calls to (the same) \link{glance} method.}
}
\description{
Use this function to test glance methods in broom tests.
Called for side effects. Will throw a useful error if
\code{glance} specification is violated. Otherwise silent.
}
\examples{

library(broom)

fit <- lm(hp ~ ., mtcars)
fit2 <- lm(hp ~ 1, mtcars)

gl <- glance(fit)
gl2 <- glance(fit2)

check_glance_outputs(gl)  # can check a single glance

# checking multiple glance outputs for same model type makes sure
# column names are consistent

check_glance_outputs(gl, gl2)

}
\seealso{
\code{\link[=check_tibble]{check_tibble()}}
}
