% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predict.R
\name{check_predict}
\alias{check_predict}
\title{Check a predict method}
\usage{
check_predict(predict_method, object, new_data = NULL, outcome = NULL,
  types = NULL)
}
\description{
Driver function that calls lots of other testsshould call twice: once with training data, once with test data
}
