% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predict_arguments.R
\name{check_predict_arguments}
\alias{check_predict_arguments}
\title{TODO: DOCUMENT!!!}
\usage{
check_predict_arguments(predict_method)
}
\arguments{
\item{tidy_method}{A tidying method. For example: \code{glance.Arima}.}

\item{strict}{Logical indicating whether the strict version of tests should be used. Defaults
to \code{TRUE}.}
}
\value{
An invisible \code{NULL}. This function should be called for side effects, not return values.
}
\description{
Call this function to perform tests. If a tests fails, an informative error
will be thrown. Otherwise silent.

Tests when \code{strict = FALSE}:
\itemize{
\item None
}

Tests when \code{strict = TRUE}:
\itemize{
\item \code{tidy_method} has a \code{conf.int} argument if it has a \code{conf.level} argument.
\item All arguments to \code{tidy_method} are listed in the \link{argument_glossary}.
}
}
\seealso{
\link{testthat}, \code{\link[testthat:expect_true]{testthat::expect_true()}}
}
