% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predict_output.R
\name{check_predict_output}
\alias{check_predict_output}
\title{Title}
\usage{
check_predict_output(predictions, passed_data, type)
}
\arguments{
\item{predictions}{TODO}

\item{passed_data}{TODO}

\item{type}{TODO}

\item{se_fit}{TODOS}
}
\value{
TODO
}
\description{
Title
}
