% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_single_augment_output}
\alias{check_single_augment_output}
\title{Check the output of an augment method}
\usage{
check_single_augment_output(au, passed_data, strict = FALSE)
}
\arguments{
\item{au}{Output from a call to \link{augment}}

\item{passed_data}{Whichever of \code{data} or \code{newdata} was passed to
\code{augment}. Should be a data frame of some sort.}
}
\description{
Do not use \code{check_single_augment_output} in broom tests.
Use \code{check_augment_function} instead, which will call this function
for many relevant inputs. As always, call for side effects. Errors
when \code{augment} specification is violated, otherwise silent.
}
\examples{

library(broom)

fit <- lm(hp ~ ., mtcars)
au <- augment(fit)
check_single_augment_output(au, mtcars)

}
