% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-tests.R
\name{check_tibble}
\alias{check_tibble}
\title{Check the output of a tidying method}
\usage{
check_tibble(output, method, check_values = FALSE,
  columns = colnames(output), strict = FALSE)
}
\arguments{
\item{output}{Object returned from \code{\link[=tidy]{tidy()}}, \code{\link[=augment]{augment()}} or \code{\link[=glance]{glance()}}.}

\item{method}{One of \code{"tidy"}, \code{"augment"} or \code{"glance"}. Determines
which set of column name checks are applied.}

\item{check_values}{Whether to check if \code{output} contains \code{NaN} or \code{Inf}.
Defaults to \code{FALSE}.}

\item{columns}{The names of the columns in the output data frame. Defaults
to the column names of \code{output}. Useful for \code{augment} when you only
want to check the new columns in the data frame, as opposed to all
columns.}
}
\description{
Do not use \code{check_tibble} in broom tests. \code{check_glance_output},
\code{check_augment_function} and \code{check_tidy_output} all call \code{check_tibble}.
Use those instead. This function is called for its side effects.
Throws a useful error if \code{output} is not a \code{tibble}. Checks that
columns in output use acceptable names as defined in the glossary.
Optionally can check for \code{NaN} or \code{Inf} values. Silent if
everything goes well.
}
\examples{

library(broom)

fit <- lm(hp ~ ., mtcars)
gl <- glance(fit)

# but don't do this, use: `check_glance_output(gl)` instead
check_tibble(gl, method = "glance", check_values = TRUE)

}
