% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Create a confusion matrix}
\usage{
confusion_matrix(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
A dataframe summing every combination
}
\description{
This function builds a confusion matrix from two logical vectors.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
confusion_matrix(t, p)
}

