% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{F1}
\alias{F1}
\title{Calculate F1 score}
\usage{
F1(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
numeric value
}
\description{
This function calculates the product of the model's recall and precision.
Lower F1 scores may indicate that a model is gaining either precision or
recall at the expense of the complementary score.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
F1(t, p)
}

