% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_rownames}
\alias{has_rownames}
\title{Check whether or not a data-frame-like object has rownames}
\usage{
has_rownames(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Logical indicating if \code{df} has rownames. If \code{df} is a tibble,
returns \code{FALSE}. If \code{df} is a data.frame, return \code{FALSE} if the rownames
are simply row numbers. If the rownames are anything other than the
return row numbers, returns \code{TRUE}.
}
\description{
Check whether or not a data-frame-like object has rownames
}
