% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltests-package.R
\docType{package}
\name{modeltests-package}
\alias{modeltests}
\alias{modeltests-package}
\title{modeltests: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
