% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{precision}
\alias{precision}
\title{Calculate precision}
\usage{
precision(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
numeric value
}
\description{
This function calcuates the fraction of the items identified by the
classifier that actually \emph{are} in the class.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
precision(t, p)
}

