% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{recall}
\alias{recall}
\title{Calculate recall}
\usage{
recall(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
numeric value
}
\description{
This function calcuates the fraction of the items \emph{within} the class that the
classifier correctly identifies.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
recall(t, p)
}

