% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity}
\usage{
specificity(truth, pred)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}
}
\value{
numeric value
}
\description{
This function calcuates the fraction of the items \emph{not} in the class that the
classifier correctly identifies
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
specificity(t, p)
}

