% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modeltests.R
\name{summary_table}
\alias{summary_table}
\title{Create a summary table of model performance measures}
\usage{
summary_table(truth, pred, name = NULL, gathered = TRUE)
}
\arguments{
\item{truth}{A logical vector of true values}

\item{pred}{A logical vector of values predicted by the model}

\item{name}{An atomic character vector that will be appended to the output
table, useful if you will be joining multipe summary tables}

\item{gathered}{A boolean value. If TRUE, returns a 'tidy' dataframe with a
column of values, and one row per measure type. If FALSE, returns a one-row
data frame with one column for each measure type.}
}
\value{
A data frame with performance measure values
}
\description{
This function returns a data frame of model accuracy, precision, recall,
specificity and F1 value.
}
\examples{
t <- c(TRUE, TRUE, FALSE, TRUE, FALSE)
p <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
summary_table(t, p, gathered = FALSE)
}

