% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dims.R
\name{check_dims}
\alias{check_dims}
\title{Check that tibble has expected dimensions.}
\usage{
check_dims(data, expected_rows = NULL, expected_cols = NULL)
}
\arguments{
\item{data}{A tibble or data frame.}

\item{expected_rows}{Expected number of rows of tibble.}

\item{expected_cols}{Expected number of columns of tibble.}
}
\description{
Check that tibble has expected dimensions.
}
\examples{

check_dims(iris, expected_rows = 150)

}
