% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltests-package.R
\docType{package}
\name{modeltests-package}
\alias{modeltests}
\alias{modeltests-package}
\title{modeltests: Testing Infrastructure for Broom Model Generics}
\description{
Provides a number of testthat tests that can be used to verify that
    tidy(), glance() and augment() methods meet specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/modeltests}
  \item Report bugs at \url{https://github.com/alexpghayes/modeltests/issues}
}

}
\keyword{internal}
