% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltests-package.R
\docType{package}
\name{modeltests-package}
\alias{modeltests}
\alias{modeltests-package}
\title{modeltests: Testing Infrastructure for Broom Model Generics}
\description{
Provides a number of testthat tests that can be used to verify that tidy(), glance() and augment() methods meet consistent specifications. This allows methods for the same generic to be spread across multiple packages, since all of those packages can make the same guarantees to users about returned objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/modeltests}
  \item Report bugs at \url{https://github.com/alexpghayes/modeltests/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

Authors:
\itemize{
  \item Simon Couch \email{simonpatrickcouch@gmail.com}
}

}
\keyword{internal}
